+++ packages/atomic/src/components.d.ts
@@ -5,7 +5,7 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { Engine, HeadlessConfigurationOptions, Result, ResultTemplateCondition } from "@coveo/headless";
+import { Engine, HeadlessConfigurationOptions, LogLevel, Result, ResultTemplateCondition } from "@coveo/headless";
 export namespace Components {
     interface AtomicBreadcrumbManager {
         "categoryDivider": string;
@@ -53,6 +53,9 @@ export namespace Components {
     interface AtomicQueryError {
     }
     interface AtomicQuerySummary {
+    }
+    interface AtomicRelevanceInspector {
+        "engine": Engine;
     }
     interface AtomicResult {
         "engine": Engine;
@@ -103,6 +106,7 @@ export namespace Components {
     }
     interface AtomicSearchInterface {
         "initialize": (options: Pick<HeadlessConfigurationOptions, 'accessToken' | 'organizationId' | 'renewAccessToken' | 'platformUrl'>) => Promise<void>;
+        "logLevel"?: LogLevel;
         "pipeline": string;
         "sample": boolean;
         "searchHub": string;
@@ -199,6 +203,12 @@ declare global {
         prototype: HTMLAtomicQuerySummaryElement;
         new (): HTMLAtomicQuerySummaryElement;
     };
+    interface HTMLAtomicRelevanceInspectorElement extends Components.AtomicRelevanceInspector, HTMLStencilElement {
+    }
+    var HTMLAtomicRelevanceInspectorElement: {
+        prototype: HTMLAtomicRelevanceInspectorElement;
+        new (): HTMLAtomicRelevanceInspectorElement;
+    };
     interface HTMLAtomicResultElement extends Components.AtomicResult, HTMLStencilElement {
     }
     var HTMLAtomicResultElement: {
@@ -274,6 +284,7 @@ declare global {
         "atomic-pager": HTMLAtomicPagerElement;
         "atomic-query-error": HTMLAtomicQueryErrorElement;
         "atomic-query-summary": HTMLAtomicQuerySummaryElement;
+        "atomic-relevance-inspector": HTMLAtomicRelevanceInspectorElement;
         "atomic-result": HTMLAtomicResultElement;
         "atomic-result-link": HTMLAtomicResultLinkElement;
         "atomic-result-list": HTMLAtomicResultListElement;
@@ -333,6 +344,9 @@ declare namespace LocalJSX {
     }
     interface AtomicQuerySummary {
     }
+    interface AtomicRelevanceInspector {
+        "engine": Engine;
+    }
     interface AtomicResult {
         "engine": Engine;
         "result": Result;
@@ -379,6 +393,7 @@ declare namespace LocalJSX {
         "numberOfSuggestions"?: number;
     }
     interface AtomicSearchInterface {
+        "logLevel"?: LogLevel;
         "pipeline"?: string;
         "sample"?: boolean;
         "searchHub"?: string;
@@ -404,6 +419,7 @@ declare namespace LocalJSX {
         "atomic-pager": AtomicPager;
         "atomic-query-error": AtomicQueryError;
         "atomic-query-summary": AtomicQuerySummary;
+        "atomic-relevance-inspector": AtomicRelevanceInspector;
         "atomic-result": AtomicResult;
         "atomic-result-link": AtomicResultLink;
         "atomic-result-list": AtomicResultList;
@@ -434,6 +450,7 @@ declare module "@stencil/core" {
             "atomic-pager": LocalJSX.AtomicPager & JSXBase.HTMLAttributes<HTMLAtomicPagerElement>;
             "atomic-query-error": LocalJSX.AtomicQueryError & JSXBase.HTMLAttributes<HTMLAtomicQueryErrorElement>;
             "atomic-query-summary": LocalJSX.AtomicQuerySummary & JSXBase.HTMLAttributes<HTMLAtomicQuerySummaryElement>;
+            "atomic-relevance-inspector": LocalJSX.AtomicRelevanceInspector & JSXBase.HTMLAttributes<HTMLAtomicRelevanceInspectorElement>;
             "atomic-result": LocalJSX.AtomicResult & JSXBase.HTMLAttributes<HTMLAtomicResultElement>;
             "atomic-result-link": LocalJSX.AtomicResultLink & JSXBase.HTMLAttributes<HTMLAtomicResultLinkElement>;
             "atomic-result-list": LocalJSX.AtomicResultList & JSXBase.HTMLAttributes<HTMLAtomicResultListElement>;