+++ packages/atomic/src/components/atomic-search-interface/atomic-search-interface.tsx
@@ -6,6 +6,7 @@ import {
   Method,
   Watch,
   Element,
+  State,
 } from '@stencil/core';
 import {
   HeadlessEngine,
@@ -15,6 +16,7 @@ import {
   HeadlessConfigurationOptions,
   AnalyticsActions,
   ConfigurationActions,
+  LogLevel,
   buildSearchParameterManager,
   buildSearchParameterSerializer,
   Unsubscribe,
@@ -31,14 +33,15 @@ export class AtomicSearchInterface {
   @Prop() sample = false;
   @Prop({reflect: true}) pipeline = 'default';
   @Prop({reflect: true}) searchHub = 'default';
+  @Prop() logLevel?: LogLevel = 'info';
   @RenderError() error?: Error;
+  @State() engine?: Engine;
 
-  private engine?: Engine;
   private unsubscribe: Unsubscribe = () => {};
   private hangingComponentsInitialization: InitializeEvent[] = [];
   private initialized = false;
 
-  componentDidLoad() {
+  componentWillLoad() {
     if (this.sample) {
       this.initialize(HeadlessEngine.getSampleConfiguration());
     }
@@ -78,6 +81,9 @@ export class AtomicSearchInterface {
     this.engine = new HeadlessEngine({
       configuration: config,
       reducers: searchAppReducers,
+      loggerOptions: {
+        level: this.logLevel,
+      },
     });
 
     this.hangingComponentsInitialization.forEach((event) =>
@@ -97,9 +103,10 @@ export class AtomicSearchInterface {
   }
 
   private initSearchParameterManager() {
-    const stateWithoutHashSign = window.location.hash.slice(1);
+    const stateWithoutHash = window.location.hash.slice(1);
+    const decodedState = decodeURIComponent(stateWithoutHash);
     const {serialize, deserialize} = buildSearchParameterSerializer();
-    const params = deserialize(stateWithoutHashSign);
+    const params = deserialize(decodedState);
 
     const manager = buildSearchParameterManager(this.engine!, {
       initialState: {parameters: params},
@@ -134,6 +141,15 @@ export class AtomicSearchInterface {
   }
 
   public render() {
-    return <slot></slot>;
+    if (!this.engine) {
+      return;
+    }
+
+    return [
+      <atomic-relevance-inspector
+        engine={this.engine}
+      ></atomic-relevance-inspector>,
+      <slot></slot>,
+    ];
   }
 }