+++ packages/headless/src/controllers/facets/facet/headless-facet.test.ts
@@ -14,10 +14,12 @@ import {buildMockFacetValue} from '../../../test/mock-facet-value';
 import {executeSearch} from '../../../features/search/search-actions';
 import {FacetRequest} from '../../../features/facets/facet-set/interfaces/request';
 import {buildMockFacetRequest} from '../../../test/mock-facet-request';
-import {buildMockFacetSearch} from '../../../test/mock-facet-search';
-import * as FacetSearch from '../facet-search/specific/headless-facet-search';
+
 import {updateFacetOptions} from '../../../features/facet-options/facet-options-actions';
 import {SearchAppState} from '../../../state/search-app-state';
+import * as FacetIdDeterminor from '../_common/facet-id-determinor';
+import {buildMockFacetSearch} from '../../../test/mock-facet-search';
+import * as FacetSearch from '../facet-search/specific/headless-facet-search';
 
 describe('facet', () => {
   const facetId = '1';
@@ -54,6 +56,17 @@ describe('facet', () => {
     expect(facet).toBeTruthy();
   });
 
+  it('it calls #determineFacetId with the correct params', () => {
+    jest.spyOn(FacetIdDeterminor, 'determineFacetId');
+
+    initFacet();
+
+    expect(FacetIdDeterminor.determineFacetId).toHaveBeenCalledWith(
+      engine,
+      options
+    );
+  });
+
   it('registers a facet with the passed options and the default values of unspecified options', () => {
     options = {
       facetId,
@@ -77,6 +90,11 @@ describe('facet', () => {
 
   it('registering a facet with #numberOfValues less than 1 throws', () => {
     options.numberOfValues = 0;
+    expect(() => initFacet()).toThrow('Check the options of buildFacet');
+  });
+
+  it('registering a facet with a space in the id throws', () => {
+    options.facetId = 'bad id';
     expect(() => initFacet()).toThrow('Check the options of buildFacet');
   });
 
@@ -418,8 +436,21 @@ describe('facet', () => {
   it('exposes a #facetSearch property', () => {
     jest.spyOn(FacetSearch, 'buildFacetSearch');
     initFacet();
-
     expect(facet.facetSearch).toBeTruthy();
-    expect(FacetSearch.buildFacetSearch).toHaveBeenCalledTimes(1);
+    expect(FacetSearch.buildFacetSearch).toHaveBeenCalled();
+  });
+
+  it('exposes a #facetSearch state', () => {
+    expect(facet.state.facetSearch).toBeTruthy();
+    expect(facet.state.facetSearch.values).toEqual([]);
+
+    const fakeResponseValue = {
+      count: 123,
+      displayValue: 'foo',
+      rawValue: 'foo',
+    };
+    engine.state.facetSearchSet[facetId].response.values = [fakeResponseValue];
+
+    expect(facet.state.facetSearch.values[0]).toMatchObject(fakeResponseValue);
   });
 });