+++ packages/headless/src/controllers/facets/facet/headless-facet.ts
@@ -7,7 +7,6 @@ import {
   updateFacetNumberOfValues,
   updateFacetIsFieldExpanded,
 } from '../../../features/facets/facet-set/facet-set-actions';
-import {randomID} from '../../../utils/utils';
 import {
   facetRequestSelector,
   facetResponseSelector,
@@ -36,6 +35,7 @@ import {validateOptions} from '../../../utils/validate-payload';
 import {defaultFacetOptions} from '../../../features/facets/facet-set/facet-set-slice';
 import {defaultFacetSearchOptions} from '../../../features/facets/facet-search-set/facet-search-reducer-helpers';
 import {FacetOptions, facetOptionsSchema} from './headless-facet-options';
+import {determineFacetId} from '../_common/facet-id-determinor';
 
 export {FacetOptions};
 export type FacetProps = {
@@ -54,15 +54,15 @@ export function buildFacet(
   const {dispatch} = engine;
   const controller = buildController(engine);
 
-  const facetId = props.options.facetId || randomID('facet');
+  const facetId = determineFacetId(engine, props.options);
   const options: Required<FacetOptions> = {
-    facetId,
     facetSearch: {...defaultFacetSearchOptions},
     ...defaultFacetOptions,
     ...props.options,
+    facetId,
   };
 
-  validateOptions(facetOptionsSchema, options, buildFacet.name);
+  validateOptions(engine, facetOptionsSchema, options, buildFacet.name);
 
   const createFacetSearch = () => {
     const {facetId, facetSearch} = options;
@@ -91,10 +91,11 @@ export function buildFacet(
   };
 
   dispatch(registerFacet(options));
-
+  const facetSearch = createFacetSearch();
+  const {state, ...restOfFacetSearch} = facetSearch;
   return {
     ...controller,
-    facetSearch: createFacetSearch(),
+    facetSearch: restOfFacetSearch,
     /**
      * Selects (deselects) the passed value if unselected (selected).
      * @param selection The facet value to select or deselect.
@@ -195,6 +196,7 @@ export function buildFacet(
 
         /** @returns `true` if fewer values can be displayed and `false` otherwise.*/
         canShowLessValues: computeCanShowLessValues(),
+        facetSearch: facetSearch.state,
       };
     },
   };