+++ packages/headless/src/controllers/search-parameter-manager/headless-search-parameter-manager.ts
@@ -1,11 +1,15 @@
 import {RecordValue, Schema} from '@coveo/bueno';
 import {Engine} from '../../app/headless-engine';
+import {getAdvancedSearchQueriesInitialState} from '../../features/advanced-search-queries/advanced-search-queries-state';
+import {getDebugInitialState} from '../../features/debug/debug-state';
+import {getPaginationInitialState} from '../../features/pagination/pagination-state';
 import {getQueryInitialState} from '../../features/query/query-state';
 import {
   restoreSearchParameters,
   SearchParameters,
 } from '../../features/search-parameters/search-parameter-actions';
 import {searchParametersDefinition} from '../../features/search-parameters/search-parameter-schema';
+import {getSortCriteriaInitialState} from '../../features/sort-criteria/sort-criteria-state';
 import {SearchParametersState} from '../../state/search-app-state';
 import {validateInitialState} from '../../utils/validate-payload';
 import {buildController} from '../controller/headless-controller';
@@ -43,6 +47,7 @@ export function buildSearchParameterManager(
   const controller = buildController(engine);
 
   validateInitialState(
+    engine,
     initialStateSchema,
     props.initialState,
     buildSearchParameterManager.name
@@ -57,6 +62,12 @@ export function buildSearchParameterManager(
       const parameters: SearchParameters = {
         ...getQ(state),
         ...getEnableQuerySyntax(state),
+        ...getAq(state),
+        ...getCq(state),
+        ...getFirstResult(state),
+        ...getNumberOfResults(state),
+        ...getSortCriteria(state),
+        ...getDebug(state),
       };
 
       return {parameters};
@@ -83,4 +94,61 @@ function getEnableQuerySyntax(state: Partial<SearchParametersState>) {
   const shouldInclude =
     enableQuerySyntax !== getQueryInitialState().enableQuerySyntax;
   return shouldInclude ? {enableQuerySyntax} : {};
+}
+
+function getAq(state: Partial<SearchParametersState>) {
+  if (state.advancedSearchQueries === undefined) {
+    return {};
+  }
+
+  const aq = state.advancedSearchQueries.aq;
+  const shouldInclude = aq !== getAdvancedSearchQueriesInitialState().aq;
+  return shouldInclude ? {aq} : {};
+}
+
+function getCq(state: Partial<SearchParametersState>) {
+  if (state.advancedSearchQueries === undefined) {
+    return {};
+  }
+
+  const cq = state.advancedSearchQueries.cq;
+  const shouldInclude = cq !== getAdvancedSearchQueriesInitialState().cq;
+  return shouldInclude ? {cq} : {};
+}
+
+function getFirstResult(state: Partial<SearchParametersState>) {
+  if (state.pagination === undefined) {
+    return {};
+  }
+
+  const firstResult = state.pagination.firstResult;
+  const shouldInclude = firstResult !== getPaginationInitialState().firstResult;
+  return shouldInclude ? {firstResult} : {};
+}
+
+function getNumberOfResults(state: Partial<SearchParametersState>) {
+  if (state.pagination === undefined) {
+    return {};
+  }
+
+  const numberOfResults = state.pagination.numberOfResults;
+  const shouldInclude =
+    numberOfResults !== getPaginationInitialState().numberOfResults;
+  return shouldInclude ? {numberOfResults} : {};
+}
+
+function getSortCriteria(state: Partial<SearchParametersState>) {
+  if (state.sortCriteria === undefined) {
+    return {};
+  }
+
+  const sortCriteria = state.sortCriteria;
+  const shouldInclude = sortCriteria !== getSortCriteriaInitialState();
+  return shouldInclude ? {sortCriteria} : {};
+}
+
+function getDebug(state: Partial<SearchParametersState>) {
+  const debug = state.debug;
+  const shouldInclude = debug !== getDebugInitialState();
+  return shouldInclude ? {debug} : {};
 }