+++ packages/headless/src/features/facets/range-facets/generic/range-facet-actions.ts
@@ -1,6 +1,6 @@
 import {RangeFacetSortCriterion} from './interfaces/request';
 import {createAction} from '@reduxjs/toolkit';
-import {validatePayloadSchema} from '../../../../utils/validate-payload';
+import {validatePayload} from '../../../../utils/validate-payload';
 import {facetIdDefinition} from '../../generic/facet-actions-validation';
 import {Value} from '@coveo/bueno';
 
@@ -12,7 +12,7 @@ import {Value} from '@coveo/bueno';
 export const updateRangeFacetSortCriterion = createAction(
   'rangeFacet/updateSortCriterion',
   (payload: {facetId: string; criterion: RangeFacetSortCriterion}) =>
-    validatePayloadSchema(payload, {
+    validatePayload(payload, {
       facetId: facetIdDefinition,
       criterion: new Value<RangeFacetSortCriterion>({required: true}),
     })