+++ packages/headless/src/features/search-parameters/search-parameter-schema.ts
@@ -1,6 +1,20 @@
-import {BooleanValue, StringValue} from '@coveo/bueno';
+import {
+  BooleanValue,
+  StringValue,
+  SchemaDefinition,
+  NumberValue,
+} from '@coveo/bueno';
+import {SearchParameters} from './search-parameter-actions';
 
-export const searchParametersDefinition = {
+export const searchParametersDefinition: SchemaDefinition<Required<
+  SearchParameters
+>> = {
   q: new StringValue(),
   enableQuerySyntax: new BooleanValue(),
+  aq: new StringValue(),
+  cq: new StringValue(),
+  firstResult: new NumberValue({min: 0}),
+  numberOfResults: new NumberValue({min: 0}),
+  sortCriteria: new StringValue(),
+  debug: new BooleanValue(),
 };